refactor(form): extract CollapsiblePipelineGroup component and move form helpers to utils
